requirements.txt
matplotlib
Pillow
numpy
os